move utils from firebase to fire test
This is more suitable since utils is basically
helper function for testing not for backend

As for the testing the utils does not
have any function with test in front of it
so its not seen as test
